requirements.txt
Pillow==6.0.0
dlib==19.17.0
numpy==1.16.2
